refactor: many API changes, see details
This is a very big refactor which will hopefully make it easier to
interact with `geo-svg`. While implementing some new advanced features I
stumbled across a hurdle again which I tripped over many times before:

The fact that the old `Svg` type captured some life times made it really
impossible to generate a svg from some temporarily available data.

This was the case since the old approach kept all the references to the
data sources as long as possible in memory to make it possible to apply
the styles at the very last second. Although this approach enables you
to render the same data with different styles, I would argue that this
was a design mistake.

The new design stream-lines the svg generation a bit more and clearly
separates defining a style for each svg part from adding the data that
should be rendered for that part. This is done with the builder pattern.

- First we build a style for some part of the svg
- After that we add the shapes that the style should be applied to
- Only then we can finish our svg document

To achieve almost-feature-pairity, there is also

- a way to combine several svg documents into one
- a way to apply some fall back general style to a svg document